+++ packages/kit/src/runtime/client/singletons.js
