+++ packages/kit/src/runtime/client/start.js
