+++ packages/kit/src/runtime/hash.js
