+++ packages/kit/src/runtime/server/page/respond_with_error.js
