+++ packages/kit/src/runtime/server/page/types.d.ts
